requirements.txt
python-telegram-bot==20.7
PyYAML==6.0.2
Jinja2==3.1.4
python-dotenv==1.0.1